+++ light-client/examples/light_client.rs
@@ -1,3 +1,12 @@
+use std::collections::HashMap;
+use std::{
+    path::{Path, PathBuf},
+    time::Duration,
+};
+
+use gumdrop::Options;
+
+use tendermint_light_client::supervisor::{Handle as _, Instance, Supervisor};
 use tendermint_light_client::{
     components::{
         clock::SystemClock,
@@ -10,17 +19,8 @@ use tendermint_light_client::{
     light_client::{self, LightClient},
     peer_list::PeerList,
     state::State,
-    store::{sled::SledStore, LightStore, VerifiedStatus},
-    supervisor::{Instance, Supervisor},
-    types::{Height, PeerId, Time, TrustThreshold},
-};
-
-use gumdrop::Options;
-
-use std::collections::HashMap;
-use std::{
-    path::{Path, PathBuf},
-    time::Duration,
+    store::{sled::SledStore, LightStore},
+    types::{Height, PeerId, Status, Time, TrustThreshold},
 };
 
 #[derive(Debug, Options)]
@@ -103,8 +103,8 @@ fn make_instance(
                 std::process::exit(1);
             });
 
-        light_store.insert(trusted_state, VerifiedStatus::Verified);
-    } else if light_store.highest(VerifiedStatus::Verified).is_none() {
+        light_store.insert(trusted_state, Status::Verified);
+    } else if light_store.highest(Status::Verified).is_none() {
         println!("[ error ] no trusted state in database, please specify a trusted header");
         std::process::exit(1);
     }
@@ -160,19 +160,19 @@ fn sync_cmd(opts: SyncOpts) {
         ProdEvidenceReporter::new(peer_addr),
     );
 
-    let mut handle = supervisor.handle();
+    let handle = supervisor.handle();
 
     std::thread::spawn(|| supervisor.run());
 
     loop {
-        handle.verify_to_highest_async(|result| match result {
+        match handle.verify_to_highest() {
             Ok(light_block) => {
-                println!("[ info  ] synced to block {}", light_block.height());
+                println!("[info] synced to block {}", light_block.height());
             }
-            Err(e) => {
-                println!("[ error ] sync failed: {}", e);
+            Err(err) => {
+                println!("[error] sync failed: {}", err);
             }
-        });
+        }
 
         std::thread::sleep(Duration::from_millis(800));
     }