+++ light-client/src/errors.rs
@@ -1,6 +1,9 @@
 //! Toplevel errors raised by the light client.
 
+use std::fmt::Debug;
+
 use anomaly::{BoxError, Context};
+use crossbeam_channel as crossbeam;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
@@ -8,7 +11,7 @@ use crate::{
     components::io::IoError,
     light_client::Options,
     predicates::errors::VerificationError,
-    types::{Height, LightBlock, PeerId},
+    types::{Height, LightBlock, PeerId, Status},
 };
 
 pub type Error = anomaly::Error<ErrorKind>;
@@ -21,6 +24,9 @@ pub enum ErrorKind {
     #[error("store error")]
     Store,
 
+    #[error("no primary")]
+    NoPrimary,
+
     #[error("no witnesses")]
     NoWitnesses,
 
@@ -31,12 +37,22 @@ pub enum ErrorKind {
     ForkDetected(Vec<PeerId>),
 
     #[error("no initial trusted state")]
-    NoInitialTrustedState,
+    NoInitialTrustedState(Status),
 
-    #[error("latest trusted state outside of trusting period")]
+    #[error("no trusted state")]
+    NoTrustedState(Status),
+
+    #[error("target height ({target_height}) is lower than trusted state ({trusted_height})")]
+    TargetLowerThanTrustedState {
+        target_height: Height,
+        trusted_height: Height,
+    },
+
+    #[error("trusted state outside of trusting period")]
     TrustedStateOutsideTrustingPeriod {
         trusted_state: Box<LightBlock>,
         options: Options,
+        status: Status,
     },
 
     #[error("bisection for target at height {0} failed when reached trusted state at height {1}")]
@@ -44,6 +60,9 @@ pub enum ErrorKind {
 
     #[error("invalid light block: {0}")]
     InvalidLightBlock(#[source] VerificationError),
+
+    #[error("internal channel disconnected")]
+    ChannelDisconnected,
 }
 
 impl ErrorKind {
@@ -62,11 +81,13 @@ pub trait ErrorExt {
     /// Whether this error means that the light block has expired,
     /// ie. it's outside of the trusting period.
     fn has_expired(&self) -> bool;
+
+    /// Whether this error means that a timeout occured when
+    /// querying a node.
+    fn is_timeout(&self) -> bool;
 }
 
 impl ErrorExt for ErrorKind {
-    /// Whether this error means that the light block
-    /// cannot be trusted w.r.t. the latest trusted state.
     fn not_enough_trust(&self) -> bool {
         if let Self::InvalidLightBlock(e) = self {
             e.not_enough_trust()
@@ -75,13 +96,32 @@ impl ErrorExt for ErrorKind {
         }
     }
 
-    /// Whether this error means that the light block has expired,
-    /// ie. it's outside of the trusting period.
     fn has_expired(&self) -> bool {
         if let Self::InvalidLightBlock(e) = self {
             e.has_expired()
         } else {
             false
         }
+    }
+
+    /// Whether this error means that a timeout occured when querying a node.
+    fn is_timeout(&self) -> bool {
+        if let Self::Io(e) = self {
+            e.is_timeout()
+        } else {
+            false
+        }
+    }
+}
+
+impl<T: Debug + Send + Sync + 'static> From<crossbeam::SendError<T>> for ErrorKind {
+    fn from(_err: crossbeam::SendError<T>) -> Self {
+        Self::ChannelDisconnected
+    }
+}
+
+impl From<crossbeam::RecvError> for ErrorKind {
+    fn from(_err: crossbeam::RecvError) -> Self {
+        Self::ChannelDisconnected
     }
 }