+++ light-client/src/store/sled.rs
@@ -5,19 +5,21 @@ use crate::{
     types::{Height, LightBlock},
 };
 
-use super::{LightStore, VerifiedStatus};
+use super::{LightStore, Status};
 use ::sled::Db as SledDb;
 
-const VERIFIED_PREFIX: &str = "light_store/verified";
 const UNVERIFIED_PREFIX: &str = "light_store/unverified";
+const VERIFIED_PREFIX: &str = "light_store/verified";
+const TRUSTED_PREFIX: &str = "light_store/trusted";
 const FAILED_PREFIX: &str = "light_store/failed";
 
 /// Persistent store backed by an on-disk `sled` database.
 #[derive(Debug, Clone)]
 pub struct SledStore {
     db: SledDb,
-    verified_db: KeyValueDb<Height, LightBlock>,
     unverified_db: KeyValueDb<Height, LightBlock>,
+    verified_db: KeyValueDb<Height, LightBlock>,
+    trusted_db: KeyValueDb<Height, LightBlock>,
     failed_db: KeyValueDb<Height, LightBlock>,
 }
 
@@ -25,53 +27,55 @@ impl SledStore {
     pub fn new(db: SledDb) -> Self {
         Self {
             db,
-            verified_db: KeyValueDb::new(VERIFIED_PREFIX),
             unverified_db: KeyValueDb::new(UNVERIFIED_PREFIX),
+            verified_db: KeyValueDb::new(VERIFIED_PREFIX),
+            trusted_db: KeyValueDb::new(TRUSTED_PREFIX),
             failed_db: KeyValueDb::new(FAILED_PREFIX),
         }
     }
 
-    fn db(&self, status: VerifiedStatus) -> &KeyValueDb<Height, LightBlock> {
+    fn db(&self, status: Status) -> &KeyValueDb<Height, LightBlock> {
         match status {
-            VerifiedStatus::Unverified => &self.unverified_db,
-            VerifiedStatus::Verified => &self.verified_db,
-            VerifiedStatus::Failed => &self.failed_db,
+            Status::Unverified => &self.unverified_db,
+            Status::Verified => &self.verified_db,
+            Status::Trusted => &self.trusted_db,
+            Status::Failed => &self.failed_db,
         }
     }
 }
 
 impl LightStore for SledStore {
-    fn get(&self, height: Height, status: VerifiedStatus) -> Option<LightBlock> {
+    fn get(&self, height: Height, status: Status) -> Option<LightBlock> {
         self.db(status).get(&self.db, &height).ok().flatten()
     }
 
-    fn update(&mut self, light_block: LightBlock, status: VerifiedStatus) {
-        let height = &light_block.height();
+    fn update(&mut self, light_block: &LightBlock, status: Status) {
+        let height = light_block.height();
 
-        for other in VerifiedStatus::iter() {
+        for other in Status::iter() {
             if status != *other {
-                self.db(*other).remove(&self.db, height).ok();
+                self.db(*other).remove(&self.db, &height).ok();
             }
         }
 
-        self.db(status).insert(&self.db, height, &light_block).ok();
+        self.db(status).insert(&self.db, &height, light_block).ok();
     }
 
-    fn insert(&mut self, light_block: LightBlock, status: VerifiedStatus) {
+    fn insert(&mut self, light_block: LightBlock, status: Status) {
         self.db(status)
             .insert(&self.db, &light_block.height(), &light_block)
             .ok();
     }
 
-    fn remove(&mut self, height: Height, status: VerifiedStatus) {
+    fn remove(&mut self, height: Height, status: Status) {
         self.db(status).remove(&self.db, &height).ok();
     }
 
-    fn highest(&self, status: VerifiedStatus) -> Option<LightBlock> {
+    fn highest(&self, status: Status) -> Option<LightBlock> {
         self.db(status).iter(&self.db).next_back()
     }
 
-    fn all(&self, status: VerifiedStatus) -> Box<dyn Iterator<Item = LightBlock>> {
+    fn all(&self, status: Status) -> Box<dyn Iterator<Item = LightBlock>> {
         Box::new(self.db(status).iter(&self.db))
     }
 }